/wwwroot/images/logo, auto apply migrations in Startup.cs and moved images controller to cms
